Move fetching device counts by group to group DAL
