+++ src/dal/group.ts
@@ -1,6 +1,6 @@
 import "server-only";
 
-import { asc, eq } from "drizzle-orm";
+import { asc, count, eq } from "drizzle-orm";
 
 import { db } from "@/db/client";
 
@@ -8,7 +8,7 @@ import { getSession } from "@/dal/session";
 
 import type { ActionResult } from "@/lib/definitions";
 
-import { deviceGroupsTable, usersTable } from "@/db/schemas";
+import { deviceGroupsTable, devicesTable, usersTable } from "@/db/schemas";
 
 export const getDeviceGroups = async (): Promise<ActionResult<Array<{ id: string; name: string }>>> => {
   try {
@@ -21,6 +21,33 @@ export const getDeviceGroups = async (): Promise<ActionResult<Array<{ id: string
       })
       .from(deviceGroupsTable)
       .innerJoin(usersTable, eq(deviceGroupsTable.userId, session.userId))
+      .orderBy(asc(deviceGroupsTable.name));
+
+    return {
+      data,
+      error: null,
+    };
+  } catch {
+    return {
+      data: null,
+      error: "Failed to load data.",
+    };
+  }
+};
+
+export const getDeviceCountsByGroup = async (): Promise<ActionResult<{ name: string; devices: number }[]>> => {
+  try {
+    const session = await getSession();
+
+    const data = await db
+      .select({
+        name: deviceGroupsTable.name,
+        devices: count(),
+      })
+      .from(deviceGroupsTable)
+      .innerJoin(devicesTable, eq(devicesTable.groupId, deviceGroupsTable.id))
+      .where(eq(deviceGroupsTable.userId, session.userId))
+      .groupBy(deviceGroupsTable.name)
       .orderBy(asc(deviceGroupsTable.name));
 
     return {
